Move mc_writer to io/mc_io.py
