+++ invisible_cities/cities/irene.py
@@ -8,8 +8,8 @@ import tables as tb
 
 from .. core.configure         import configure
 from .. core.system_of_units_c import units
-from .. core.mctrk_functions   import mc_track_writer
 
+from .. io.mc_io              import mc_track_writer
 from .. io.pmap_io            import pmap_writer
 from .. io.run_and_event_io   import run_and_event_writer
 from .. reco                  import tbl_functions as tbl
